+++ http-client/src/main/scala/com/twitter/finatra/httpclient/modules/HttpClientModule.scala
@@ -4,7 +4,7 @@ import com.google.inject.Provides
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.finagle.{Http, Resolvers, Service}
-import com.twitter.finatra.httpclient.{HttpClient, RichHttpClient}
+import com.twitter.finatra.httpclient.HttpClient
 import com.twitter.finatra.jackson.ScalaObjectMapper
 import com.twitter.inject.Injector
 import javax.inject.Singleton
@@ -49,19 +49,4 @@ abstract class HttpClientModule extends HttpClientModuleTrait {
     injector: Injector,
     statsReceiver: StatsReceiver
   ): Service[Request, Response] = newService(injector, statsReceiver)
-
-  // retained for binary compatibility
-  @deprecated(
-    "Use provideHttpService(injector, statsReceiver). This method no longer binds " +
-      "to the Injector's object graph.",
-    "07-09-2019")
-  final def provideHttpService: Service[Request, Response] = {
-    sslHostname match {
-      case Some(ssl) =>
-        RichHttpClient.newSslClientService(sslHostname = ssl, dest = dest)
-      case _ =>
-        RichHttpClient.newClientService(dest = dest)
-    }
-  }
-
 }